building ClientGame from packet succeeds!

+++ src/aoop/asteroids/model/Spaceship.java
@@ -242,13 +242,13 @@ public class Spaceship extends GameObject
 	}
 	
 	public static Spaceship fromJSON(JSONArray json){
-		int x = (int) json.get(0);
-		int y = (int) json.get(1);
+		int x = ((Long) json.get(0)).intValue();
+		int y = ((Long) json.get(1)).intValue();
 		double velocityX = (double) json.get(2);
 		double velocityY = (double) json.get(3);
 		double direction = (double) json.get(4);
-		boolean isAccelerating = ((int) json.get(5)) == 1;
-		int score = (int) json.get(6);
+		boolean isAccelerating = ((long) json.get(5)) == 1;
+		int score = ((Long) json.get(6)).intValue();
 		return new Spaceship(new Point(x,y),velocityX, velocityY, 15, direction, isAccelerating, score);
 	}
